issue #3641 fix calculation of next service appointment

+++ rental_product_instance/models/instance_operating_data.py
@@ -14,7 +14,7 @@ class InstanceOperatingData(models.Model):
         required=True,
     )
 
-    operating_data = fields.Integer(
+    operating_data = fields.Float(
         string='Operating Data',
         required=True,
     )
